platform: Move stm32f4 boards to platform/stm32/f4
