+++ platform/stm32/f4/nucleo_f429zi/arch.c
