+++ platform/stm32/f4/nucleo_f429zi/stm32f4xx_hal_conf.h
