+++ platform/stm32/f4/stm32f4_discovery/arch.c
