+++ platform/stm32/f4/stm32f4_discovery/stm32f4xx_hal_conf.h
