+++ src/drivers/spi/stm32/stm32_spi1.c
@@ -14,10 +14,10 @@
 
 #if defined(STM32F407xx)
 #include <framework/mod/options.h>
-#include <config/third_party/bsp/stmf4cube/stm32f4_discovery/arch.h>
+#include <config/platform/stm32/f4/stm32f4_discovery/arch.h>
 #elif defined(STM32L476xx)
 #include <framework/mod/options.h>
-#include <config/third_party/bsp/stml4cube/nucleo_l476rg/arch.h>
+#include <config/platform/stm32/f4/nucleo_l476rg/arch.h>
 #endif
 
 #include "stm32_spi.h"
@@ -34,9 +34,9 @@ static struct stm32_spi stm32_spi1 = {
 static int stm32_spi1_init(void) {
 	GPIO_InitTypeDef  GPIO_InitStruct;
 
-#if defined(STM32F407xx) && !OPTION_MODULE_GET(third_party__bsp__stmf4cube__stm32f4_discovery__arch,BOOLEAN,errata_spi_wrong_last_bit)
+#if defined(STM32F407xx) && !OPTION_MODULE_GET(platform__stm32__f4__stm32f4_discovery__arch,BOOLEAN,errata_spi_wrong_last_bit)
 	#error errata_spi_wrong_last_bit is not enabled for SPI1! \
-	       Please, enable this option in third_party.bsp.stmf4cube.stm32f4_discovery.arch
+	       Please, enable this option in platform.stm32.f4.stm32f4_discovery.arch
 #endif
 
 	spi1_enable_gpio_clocks();
